refactor: group auth and payment routes
