+++ backend/src/routes/index.js
@@ -3,66 +3,16 @@ const router = express.Router();
 
 router.use('/api/health', require('./health.routes'));
 router.use('/api/cache', require('./cache.routes'));
-router.use('/api/auth', require('../modules/auth/routes/auth.routes'));
-router.use('/api/users', require('../modules/users/user.routes'));
-router.use('/api/verify', require('../modules/verify/verify.routes'));
-router.use('/api/license', require('../modules/license/license.routes'));
-router.use(
-  '/api/certificates',
-  require('../modules/users/tutorials/certificate/certificatePublic.routes')
-);
-router.use(
-  '/api/certificates/admin',
-  require('../modules/users/tutorials/certificate/certificateAdmin.routes')
-);
-router.use('/api/certificate-templates', require('../modules/certificateTemplates/certificateTemplates.routes'));
+// grouped routers
+router.use(require('./auth'));
+router.use(require('./payments'));
 router.use('/api/bookings/admin', require('../modules/bookings/bookings.routes'));
 router.use('/api/bookings/student', require('../modules/bookings/student.routes'));
 router.use('/api/bookings/instructor', require('../modules/bookings/instructor.routes'));
 router.use('/api/community/admin', require('../modules/community/admin/admin.routes'));
 router.use('/api/community', require('../modules/community/public/public.routes'));
 router.use('/api/related-questions', require('../modules/community/public/relatedQuestions.routes'));
-router.use('/api/roles', require('../modules/roles/roles.routes'));
-router.use('/api/plans', require('../modules/plans/plans.routes'));
-router.use(
-  '/api/user-subscriptions',
-  require('../modules/subscriptions/subscriptions.routes')
-);
-// Register admin routes before public routes to prevent public routes from catching
-// requests intended for admin endpoints such as "/api/payment-methods/admin".
-router.use(
-  '/api/payment-methods/admin',
-  require('../modules/paymentMethods/paymentMethods.routes')
-);
-router.use(
-  '/api/payment-methods',
-  require('../modules/paymentMethods/paymentMethods.public.routes')
-);
-router.use('/api/payments/student', require('../modules/payments/student.routes'));
-router.use('/api/payments/bank', require('../modules/payments/bank.routes'));
-router.use('/api/payments/crypto', require('../modules/payments/crypto.routes'));
-router.use('/api/payments/coinbase', require('../modules/payments/coinbase.routes'));
-// Alias for NOWPayments crypto gateway
-router.use(
-  '/api/payments/nowpayments',
-  require('../modules/payments/crypto.routes')
-);
-// PayPal order creation and callback
-router.use(
-  '/api/payments/paypal',
-  require('../modules/payments/paypal.routes')
-);
-router.use('/api/payments/stripe', require('../modules/payments/stripe.routes'));
-router.use('/api/payments/admin', require('../modules/payments/payments.routes'));
-router.use('/api/invoices/admin', require('../modules/invoices/invoices.routes'));
-router.use('/api/invoices/student', require('../modules/invoices/student.routes'));
-router.use('/api/invoices/instructor', require('../modules/invoices/instructor.routes'));
-router.use(
-  '/api/admin/payments/bank',
-  require('../modules/payments/bank.admin.routes')
-);
 router.use('/api/admin/cache', require('./cache.routes'));
-router.use('/api/payments/config', require('../modules/paymentConfig/paymentConfig.routes'));
 router.use('/api/messages/config', require('../modules/messagesConfig/messagesConfig.routes'));
 router.use('/api/social-login/config', require('../modules/socialLoginConfig/socialLoginConfig.routes'));
 router.use('/api/app-config', require('../modules/appConfig/appConfig.routes'));
@@ -76,7 +26,6 @@ router.use('/api/contact', require('../modules/contact/contact.routes'));
 router.use('/api/seo-config', require('../modules/seoConfig/seoConfig.routes'));
 router.use('/api/popup-announcements', require('../modules/popupAnnouncements/popupAnnouncements.routes'));
 router.use('/api/policies', require('../modules/policies/policies.routes'));
-router.use('/api/payouts', require('../modules/payouts/payouts.routes'));
 router.use('/api/ads', require('../modules/ads/ads.routes'));
 router.use('/api/coupons', require('../modules/coupons/coupons.routes'));
 router.use('/api/groups', require('../modules/groups/groups.routes'));
